move refactored modal and better naming
